Plans out what needs to be done
