+++ src/proposers.rs
@@ -1,11 +1,13 @@
 use std::fmt;
-use crate::message::Message;
+use crate::messages::Message;
+use crate::messages::Status;
 use crate::acceptors::Acceptor;
 
 pub struct Proposer{
     id: u32,
     val: u32,
     num_acceptors: u32,
+    status: Status,
 
 }
 
@@ -23,6 +25,10 @@ impl Proposer{
         self.val
     }
 
+    pub fn status(&self) -> &Status{
+        &self.status
+    }
+
     pub fn set_num_acceptors(&mut self, num: u32){
         self.num_acceptors = num;
     }
@@ -35,20 +41,14 @@ impl Proposer{
         self.val = val;
     }
 
-    pub fn phase_one(&mut self, list: &mut [Acceptor]){
-        println!("Phase 1");
-        for acc in list {
-            println!("{}",acc);
-        }
-
+    pub fn set_status(&mut self, status: Status){
+        self.status = status;
     }
 
-    pub fn phase_two(&mut self, list: &mut [Acceptor]){
-        println!("Phase 2");
-        for acc in list {
-            println!("{}",acc);
-        }
-
+    pub fn run(&mut self, list: &Vec<Acceptor>){
+        // Check status of proposer
+            // If status is Active and value > 0 is set, send propose messages out
+            // Otherwise, deal with messages in the message queue and perform actions based on them 
     }
 }
 
@@ -58,13 +58,14 @@ impl Default for Proposer{
             id: 0,
             val: 0,
             num_acceptors: 0,
+            status: Status::Active
         }
     }
 }
 
 impl fmt::Display for Proposer{
     fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        write!(f, "Proposer => num_acceptors: {}, id: {}, val: {}", self.num_acceptors, self.id, self.val)
+        write!(f, "Proposer => num_acceptors: {}, id: {}, val: {}, status: {}", self.num_acceptors, self.id, self.val, self.status)
     }
 
 }
